fix project member issue

+++ src/main/java/com/sehoprojectmanagerapi/repository/project/projectmember/ProjectMemberRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProjectMemberRepository extends JpaRepository<ProjectMember, Integer> {
+public interface ProjectMemberRepository extends JpaRepository<ProjectMember, Long> {
     List<ProjectMember> findByUserId(Long userId);
 
     Optional<ProjectMember> findByUserIdAndProjectId(Long userId, Long projectId);
